Project/back/requirements.txt
Django==5.1.6
djangorestframework==3.14.0
django-cors-headers==4.3.1 